% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{bifplot}
\alias{bifplot}
\title{Plot the bifurcation diagram}
\usage{
bifplot(X, cparam, showmean = FALSE, ...)
}
\arguments{
\item{X}{The output of solve_in_range(), a matrix}

\item{cparam}{The control parameter. It should be a sequence.}

\item{showmean}{Should the mean state of potentially many variables in X be drawn?}

\item{...}{Passed to matplot(); col = 2 is used for the mean state if showmean = TRUE}
}
\description{
Given the output of localsolver::solve_in_range(), plot the bifurcation diagram of the system as a function of a control parameter.
}
\details{
This is meant as a quick diagnostic plot.
}
\examples{
library(deSolve)
library(localsolver)
rs <- .saddlenode$rs
x <- solve_in_range(rs, "r", saddlenode, .saddlenode$xinit, .saddlenode)
bifplot(x, rs)

library(igraph)
g <- sample_pa(25, m = 2, directed = FALSE, start.graph = make_full_graph(3))
A <- as_adj(g, "both", sparse = FALSE)
N <- vcount(g)
params <- c(list(A = A), .doublewell)
control <- list(times = seq(0, 1, length.out = 100))
X <- solve_in_range(params$Ds, "D", doublewell, rep(params$xinit.low, N), params, control)
bifplot(X, params$Ds, showmean = TRUE, col = 1)
}
