% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games.R
\name{death_birth}
\alias{death_birth}
\title{Death-birth updating}
\usage{
death_birth(u, g, params)
}
\arguments{
\item{u}{The node. Can be an integer node index or an igraph.vs.}

\item{g}{The network, an igraph object.}

\item{params}{A list of parameters. Must include eta, b, and c.}
}
\value{
An integer vector of length 1, a node state.
}
\description{
Performs the death-birth updating rule on a network.
}
\details{
The `params` list should include, at a minimum: eta, the selection strength; b, the benefit in the donation game; and c, the cost in the donation game.

This function takes, as its main input, a selected node, u. Ordinarily this will be a randomly selected node, but the function doesn't require it. Then, a neighbor v of node u is selected at random, proportional to the reproductive rates of u's neighbors. This function returns the state of v, which should ordinarily be assigned to u, completing the death-birth cycle.
}
\examples{
library(igraph)
library(sdn)

g <- make_lattice(length = 3, dim = 1)
E(g)$weight <- c(1, 2) # so can tell the difference
V(g)$state <- c(0, 0, 1)
params <- list(b = 2, c = 1, eta = 0.025)

                                        # test for 1 nbr
u <- 1

res <- replicate(100, death_birth(u, g, params)) # s.b. 0
stopifnot(sum(res) == 0)

                                        # test for 2 nbr
u <- 2
res <- replicate(100, death_birth(u, g, params))
stopifnot(table(res)[1] > 33.3 - 6 & table(res)[1] < 33.3 + 6)
}
\references{
Allen et al. 2017, doi:10.1038/nature21723.
}
