% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games.R
\name{ede}
\alias{ede}
\title{Evolutionary games on networks}
\usage{
ede(statevector, g, simtime, update, params, keep_history = FALSE)
}
\arguments{
\item{statevector}{A vector of length N. Should be all 0 or 1.}

\item{g}{A network, should be an igraph object.}

\item{simtime}{Number of simulation steps, an integer.}

\item{update}{The update rule, a function.}

\item{params}{A list of parameters. Must include eta, b, and c.}

\item{keep_history}{Logical.}
}
\value{
The final state vector of the network, or, if keep_history=TRUE, a matrix with `simtime` rows and `length(statevector)` columns.
}
\description{
The goal of this function is to perform arbitrary games on a network of N nodes.
}
\details{
The `params` list should include, at a minimum: eta, the selection strength; b, the benefit in the donation game; and c, the cost in the donation game.

Given a vector of node states `statevector`, a network `g`, the number of simulation iterations to perform `simtime`, an update rule `update`, and a list of parameters `params`, simulates an evolutionary game following the methods described in Allen et al (2017). The simulation will stop when `simtime` is reached or if the state vector is either all 0 or all 1. 

The name for this function follows the pattern ode (ordinary differential equations), sde (stochastic differential equations), ede (evolutionary...), but the 'de' doesn't mean anything.
}
\examples{
library(igraph)
library(sdn)

g <- make_lattice(length = 3, dim = 1)
E(g)$weight <- c(1, 2)

statevector <- c(0, 0, 1)
simtime <- 10
params <- list(b = 2, c = 1, eta = 0.025)

res <- ede(statevector, g, simtime, death_birth, params, TRUE)
res # results may vary
}
\references{
Allen et al. 2017, doi:10.1038/nature21723.
}
