% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barzon-dynamics.R
\docType{data}
\name{neuronal}
\alias{neuronal}
\alias{.neuronal}
\title{Neuronal}
\format{
An object of class \code{list} of length 4.
}
\usage{
neuronal(t, x, params)

.neuronal
}
\arguments{
\item{t}{An arbitrary time.}

\item{x}{The current state of the system. When N > 1, x is a vector with x_i representing the current state of the i'th node.}

\item{params}{A list of model parameters.}
}
\value{
A vector of derivatives.
}
\description{
A simple model of neuron activity(?).
}
\details{
A simple model in deSolve's format.
}
\references{
Hens
}
\keyword{datasets}
