% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barzon-dynamics.R
\docType{data}
\name{population}
\alias{population}
\alias{.population}
\title{Population}
\format{
An object of class \code{list} of length 5.
}
\usage{
population(t, x, params)

.population
}
\arguments{
\item{t}{An arbitrary time.}

\item{x}{The current state of the system. When N > 1, x is a vector with x_i representing the current state of the i'th node.}

\item{params}{A list of model parameters.}
}
\value{
A vector of derivatives.
}
\description{
A simple population model.
}
\details{
A simple model in deSolve's format.
}
\references{
Harush
}
\keyword{datasets}
