% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal-forms.R
\docType{data}
\name{saddlenode}
\alias{saddlenode}
\alias{.saddlenode}
\title{Saddle-node bifurcations}
\format{
An object of class \code{list} of length 3.
}
\usage{
.saddlenode

saddlenode(t, x, params)
}
\arguments{
\item{r}{The control parameter.}
}
\description{
The normal form of the saddle-node bifurcation.
}
\examples{
library(deSolve)
library(sdn)
rs <- .saddlenode$rs
x <- solve_in_range(rs, "r", saddlenode, .saddlenode$xinit, .saddlenode)
xlim <- range(rs) + c(-0.1, 0.1)
ylim <- c(min(x), max(c(max(x), 0.1)))
plot(rs, x, type = "l", xlab = "r", ylab = "x", col = 2, lwd = 2,
xlim = xlim, ylim = ylim, xaxs = "i", yaxs = "i")
abline(h = 0)
abline(v = 0)
}
\references{
https://en.wikipedia.org/wiki/Saddle-node_bifurcation
}
\keyword{datasets}
