% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barzon-dynamics.R
\docType{data}
\name{synchronization}
\alias{synchronization}
\alias{.synchronization}
\title{Synchronization}
\format{
An object of class \code{list} of length 3.
}
\usage{
synchronization(t, x, params)

.synchronization
}
\arguments{
\item{t}{An arbitrary time.}

\item{x}{The current state of the system. When N > 1, x is a vector with x_i representing the current state of the i'th node.}

\item{params}{A list of model parameters.}
}
\value{
A vector of derivatives.
}
\description{
A simple Kuramoto model.
}
\details{
A simple model in deSolve's format. Make sure to change the default values of omega and xinit in .synchronization to observe non-constant behavior.
}
\references{
Barzon only?
}
\keyword{datasets}
