% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{time_ev}
\alias{time_ev}
\title{Plot the time evolution of a system}
\usage{
time_ev(X, showmean = FALSE, ...)
}
\arguments{
\item{X}{The output of ode() or sde()}

\item{showmean}{Should the mean state of potentially many variables in X be drawn?}

\item{...}{Passed to matplot() and, if showmean = TRUE, lines()}
}
\description{
Given the output of deSolve::ode() or sdn::sde(), plot the time evolution of the system.
}
\details{
This is meant as a quick diagnostic plots. At least one consequence of this is that `col` is hard-coded inside the function...
}
\examples{
library(graphics)
library(deSolve)
library(sdn)

control <- list(times = seq(0, 20, by = 0.1))
x <- ode(.saddlenode$xinit, control$times, saddlenode, list(r = 0))
time_ev(x)
abline(h = 0, lty = 2)

library(igraph)
g <- sample_pa(25, m = 2, directed = FALSE, start.graph = make_full_graph(3))
A <- as_adj(g, "both", sparse = FALSE)
N <- vcount(g)
params <- c(list(A = A), .doublewell)
control <- list(times = seq(0, 1, length.out = 100))
X <- ode(rep(.doublewell$xinit.low, N), control$times, doublewell, params)
time_ev(X, showmean = TRUE)
}
