% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal-forms.R
\docType{data}
\name{transcritical}
\alias{transcritical}
\alias{.transcritical}
\title{Transcritical bifurcations}
\format{
An object of class \code{list} of length 4.
}
\usage{
.transcritical

transcritical(t, x, params)
}
\arguments{
\item{r}{The control parameter.}
}
\description{
The normal form of the transcritical bifurcation
}
\examples{
library(deSolve)
library(localsolver)
rs <- .transcritical$rs
x <- solve_in_range(rs, "r", transcritical, .transcritical$xinit, .transcritical)
plot(rs, x, type = "l", xlab = "r", ylab = "x", col = 2, lwd = 2, xaxs = "i")
abline(h = 0)
abline(v = 0)
}
\references{
https://en.wikipedia.org/wiki/Transcritical_bifurcation
}
\keyword{datasets}
